Replace problematic AuthControllerTest with working AuthControllerIntegrationTest
